Initial commit
This is the "polished" version of the code, not each iteration
as I improved things from taking 4 seconds down to .668.

I took notes as I progressed, the idea is two articles, one
discussing how long it took to do this, and what loccount measures...

and the methods I applied to gradually speed it up to where it
can't be made any faster

and then to speeding up loccount in go also.